==> pure-sagemaker-workshop/5.BatchTranform.ipynb <==
## Batch Transform

Now we are going to use "today's" features to create predictions, that the business unit is going to use as an input for promotions. 

For this, we are going to deploy the model created on the best training job from the hyperparameter tunning job and use the resulting endpoint for inference. 

import sagemaker
import boto3
from sagemaker.estimator import Estimator
from sagemaker.tuner import HyperparameterTuner
import numpy as np                                # For matrix operations and numerical processing
import pandas as pd                               # For munging tabular data
import os 
import time
from sagemaker.predictor import csv_serializer,RealTimePredictor

# take the best training job from notebook #3
best_training_job = 'forecast-vtas-190702-1408-004-46a4c8e2'

df = pd.read_csv('to_predict.csv',header=None)

df.shape

id_reseller = pd.read_csv('id_reseller_to_predict.csv',header=None)[0]

id_reseller.shape

model = Estimator.attach(best_training_job)

model_predictor = model.deploy(initial_instance_count=1,
                            instance_type='ml.m4.xlarge')

#model_predictor = RealTimePredictor('forecast-vtas-190528-2006-003-486b05ad')

model_predictor.content_type = 'text/csv'
model_predictor.serializer = csv_serializer
model_predictor.deserializer = None

def predict(data, rows=500):
    split_array = np.array_split(data, int(data.shape[0] / float(rows) + 1))
    predictions = ''
    for array in split_array:
        predictions = ','.join([predictions, model_predictor.predict(array).decode('utf-8')])

    return np.fromstring(predictions[1:], sep=',')

predictions = predict(df.as_matrix())

predictions.shape

df_predictions  = pd.DataFrame({'id_reseller':id_reseller,'prediction':predictions})

df_predictions.head()

Finally we upload predictions to S3

role = sagemaker.get_execution_role()
bucket = sagemaker.Session().default_bucket()                     
prefix = 'predictions'

df_predictions.to_csv('predictions.csv',index=False)

boto3.Session().resource('s3').Bucket(bucket).Object(os.path.join(prefix, 'predictions.csv')).upload_file('predictions.csv')
